pyproject.toml
[tool.poetry]
name = "juliet"
version = "1.0.1"
description = "A simple IRC relay bot for radio networks."
authors = ["W0JHX <jheddings@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
pyserial = "^3.5"
pyyaml = "^6.0.1"
irc = "^20.3.0"

[tool.poetry.group.dev.dependencies]
GitPython = "^3.1.36"
pre-commit = "^3.4.0"
pytest = "^7.4.2"
coverage = "^7.3.1"

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "I",   # isort
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "G",   # flake8-logging-format
    "RET", # flake8-return
    "N",   # pep8-naming
    "ERA", # eradicate
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line-too-long
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["juliet"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
